+++ src/components/inputs/InputTextField.js
@@ -1,46 +1,23 @@
 import React, { useState, useEffect, useRef } from 'react';
 import Autocomplete from '@mui/material/Autocomplete';
 import TextField from '@mui/material/TextField';
-import axios from 'axios';
 
-export default function InputTextField({ onProtocolSelect }) {
-  const [protocols, setProtocols] = useState([]);
+export default function InputTextField({ onProtocolSelect,protocols }) {
+  const [inputValue, setInputValue] = useState(''); // État pour la valeur du champ de texte
   const hasFetched = useRef(false); // Réf pour suivre si le fetch a été fait
+  const [selectedValue, setSelectedValue] = useState(null); // Valeur sélectionnée
 
-  // Fetch protocols when component mounts
-  useEffect(() => {
-    fetchProtocols();
-  }, []);
-
-  const fetchProtocols = () => {
-    // Ne fetch que si pas déjà fait
-    if (!hasFetched.current) {
-      hasFetched.current = true; // Marquer comme fetché
-      
-      axios.get('/api/protocolsName')
-        .then(response => {
-          const data = response.data?.listNames || [];
-          setProtocols(Array.isArray(data) ? data : []);
-        })
-        .catch(error => {
-          console.error("Fetch error:", error);
-          setProtocols(["Bonjour", "Banane", "haha"]);
-        });
-    }
-  };
-
-  // Fetch au montage du composant
-  useEffect(() => {
-    fetchProtocols();
-  }, []);
 
   const handleChange = (event, newValue) => {
     if (newValue) {
       onProtocolSelect(newValue);
       // Option: retirer le protocole sélectionné
-      setProtocols(prev => prev.filter(p => p !== newValue));
     }
   };
+    
+
+
+    
 
   return (
     <Autocomplete
@@ -49,6 +26,10 @@ export default function InputTextField({ onProtocolSelect }) {
       sx={{ width: 300 }}
       onChange={handleChange}
       autoHighlight={true}
+      inputValue={inputValue} 
+      onInputChange={(event, newInputValue) => {
+        setInputValue(newInputValue); // Met à jour la valeur du champ
+      }}
       renderInput={(params) => (
         <TextField
           {...params}
