+++ src/components/inputs/InputValidateButtons.js
@@ -6,16 +6,21 @@ import axios from 'axios';
 import "./InputValidateButtons.css";
 import { Grid2 } from '@mui/material';
 
-export default function InputValidateButtons({ protocol, onApiResult }) {
+export default function InputValidateButtons({ protocol, onApiResult,listProtocols }) {
   const [showCriteria, setShowCriteria] = useState(false);
 
   useEffect(() => {
     const handleKeyDown = (event) => {
       if (event.key === 'Enter') {
-        handleSubmit();
+        if (!protocol || protocol.trim() === '') {
+          return;    
+        }
+        else{
+          handleSubmit();
+        }
+
       }
     };
-
     window.addEventListener('keydown', handleKeyDown);
     return () => {
       window.removeEventListener('keydown', handleKeyDown);
@@ -33,6 +38,7 @@ export default function InputValidateButtons({ protocol, onApiResult }) {
           console.log(response.data);
           onApiResult(response.data);
           setShowCriteria(true); // Show the criteria after successful response
+          listProtocols(prev => prev.filter(p => p !== protocol));
       })
       .catch(error => {
         console.error("API Error:", error);
